+++ components/layout/DropZone.tsx
@@ -1,6 +1,6 @@
 import React, { Children, PropsWithChildren } from 'react';
 import { LuX } from 'react-icons/lu';
-import { DropZoneProvider, useDropZone } from '@/components/dropzone/DropZoneContext';
+import { DropZoneProvider, useDropZone } from '@/components/layout/DropZoneContext';
 import { Button } from '@/components/ui/button';
 import { cn } from '@/lib/utils';
